PROCESSING-2822 : New interface created for RequestClient, and a default implementation using Google Http api. Added new configuration param to instantiate the implementation of RequestClient that the user wishes.
